requirements.txt
Pillow>=10.0.0
requests>=2.28.0
PyYAML>=6.0
pydub>=0.25.1
tqdm>=4.66.0
